Move identity models to models.py and make migrations file
